move nano related functions to @types
